Safari-609.1.11.1/LayoutTests/platform/wincairo/plugins/plugin-initiate-popup-window-expected.txt

CONSOLE MESSAGE: line 17: window count: 1
MISSING PLUGIN BUTTON PRESSED
FAIL: Timed out waiting for notifyDone to be called


Specify a script and a mouse/keyboard event to the plugin. The specified script will be evaluated in the browser when the specified event is received by the plugin. The test is for bug https://bugs.webkit.org/show_bug.cgi?id=41292.
Opening window by mouse down is FAILED
Opening window by key down is FAILED
